% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMHMMR.R
\docType{class}
\name{ModelMHMMR-class}
\alias{ModelMHMMR-class}
\alias{ModelMHMMR}
\title{A Reference Class which represents a fitted MHMMR model.}
\description{
ModelMHMMR represents an estimated MHMMR model.
}
\section{Fields}{

\describe{
\item{\code{paramMHMMR}}{A \link{ParamMHMMR} object. It contains the estimated
values of the parameters.}

\item{\code{statMHMMR}}{A \link{StatMHMMR} object. It contains all the statistics
associated to the MHMMR model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("predicted", "filtered", "smoothed", "regressors"))}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"predicted" = } Predicted time series and predicted
      regime probabilities.
      \item \code{"filtered" = } Filtered time series and filtering
      regime probabilities.
      \item \code{"smoothed" = } Smoothed time series, and
      segmentation.
      \item \code{"regressors" = } Polynomial regression components.
    }
  }
}
By default, all the above graphs are produced.}

\item{\code{summary()}}{Summary method.}
}}

\seealso{
\link{ParamMHMMR}, \link{StatMHMMR}
}
